% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_params}
\alias{make_params}
\title{Create list of structured parameters}
\usage{
make_params(trainable_struct = FALSE, pretrain = TRUE, lambdas = NULL)
}
\arguments{
\item{trainable_struct}{a logical indicating whether the structured
network part should be trained or not. Only use FALSE if model is pretrained
(pretain = TRUE).}

\item{pretrain}{a logical indicating whether the structured network should be 
pretrained using a PAMM.}

\item{lambdas}{a list of lambda values for the structured subnetwork.
Each specified effect can receive a lambda value that results in a 
L2 penalty on structured coefficient and a prefactor for splines.
Defaults to 0 for linear effects and 1 for splines meaning no regularization
for linear effects and the default penalization suggested by mgcv for splines.
For linear effects the lambda value is simply the regularization value
while for splines a value >1 means a stronger regularization than the
default while <1 results in smaller regularization.}
}
\value{
a list that is complaint with the <param> argument of the
<deeppamm> method <make_net>.
}
\description{
This function creates a list of parameters used for the structured part
of the deeppamm object.
The list is passed towards the $make_model method.
}
